examples/python/requirements.txt
nexlayer-python>=1.0.0
requests>=2.28.0
pyyaml>=6.0.0 